Initial implementation of untexture sprites

+++ render/camera.cpp
@@ -18,15 +18,15 @@ void rotate(camera_state_t *camera, float amount)
     rot(&camera->planeX, &camera->planeY, cosrs, sinrs);
 }
 
-uint8_t sample_world_map(map_t *map, int x, int y)
+uint8_t sample_world_map(map_t *map, float x, float y)
 {
-    return map->wall_map[x + y * map->map_width];
+    return map->wall_map[int(x) + int(y) * map->map_width];
 }
 
 void move(camera_state_t *camera, map_t *map, float amount)
 {
-    if (sample_world_map(map, int(camera->posX + camera->dirX * amount), int(camera->posY)) == 0)
+    if (sample_world_map(map, camera->posX + camera->dirX * amount, camera->posY) == 0)
         camera->posX += camera->dirX * amount;
-    if (sample_world_map(map, int(camera->posX), int(camera->posY + camera->dirY * amount)) == 0)
+    if (sample_world_map(map, camera->posX, camera->posY + camera->dirY * amount) == 0)
         camera->posY += camera->dirY * amount;
 }
